Define a macro PRINTF_F in dd_share.h, which indicates printf-like functions, so that GCC can check their parameters. If GCC is not the compiler, PRINTF_F defines to nothing.
Make sure the parameters match the format string for all printf-like functions in the engine and games.

Thanks to chasonr for the patch.

+++ doomsday/engine/portable/src/dd_main.c
@@ -623,8 +623,9 @@ static int DD_StartupWorker(void *parm)
         {
             strncpy(buff, lumpinfo[p].name, 8);
             buff[8] = 0;
-            printf("%04i - %-8s (hndl: %p, pos: %i, size: %i)\n", p, buff,
-                   lumpinfo[p].handle, lumpinfo[p].position, lumpinfo[p].size);
+            printf("%04i - %-8s (hndl: %p, pos: %i, size: %lu)\n", p, buff,
+                   lumpinfo[p].handle, lumpinfo[p].position, 
+                   (unsigned long) lumpinfo[p].size);
         }
         Con_Error("---End of lumps---\n");
     }
